Brian

+++ phase3/libuser.h
@@ -13,9 +13,9 @@ extern void Terminate(long status);
 extern void GetTimeofDay(int *tod);
 extern void CPUTime(int *cpu);
 extern void GetPID(int *pid);
-extern int  SemCreate(int value, int *semaphore);
-extern int  SemP(int semaphore);
-extern int  SemV(int semaphore);
+extern int  SemCreate(long value, int *semaphore);
+extern int  SemP(long semaphore);
+extern int  SemV(long semaphore);
 extern int  SemFree(int semaphore);
 
 #endif
